% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{array_2D_from_df}
\alias{array_2D_from_df}
\title{Convert landmark table to 2D array}
\usage{
array_2D_from_df(
  df,
  specimen_column = "specimen",
  LM_column = "landmark",
  X_column = "X",
  Y_column = "Y",
  Z_column = "Z"
)
}
\arguments{
\item{df}{A \code{data.frame} with columns containing specimen names, landmark
names, X-coordinates, Y-coordinates, and Z-coordinates.}

\item{specimen_column}{Character string with the name of the specimen column.
Default: \code{"specimen"}.}

\item{LM_column}{Character string with the name of the specimen column.
Default: \code{"landmark"}.}

\item{X_column}{Character string with the name of the specimen column.
Default: \code{"X"}.}

\item{Y_column}{Character string with the name of the specimen column.
Default: \code{"Y"}.}

\item{Z_column}{Character string with the name of the specimen column.
Default: \code{"Z"}.}
}
\value{
Returns a tibble containing a \code{specimen} column and one column per
landmark and coordinate, e.g. \code{landmark1_X}, \code{landmark1_Y}, \code{landmark1_Z},
\code{landmark2_X}, \code{landmark2_Y}, \code{landmark2_Z}, \link{...}.
}
\description{
Converts a table with landmark X-, Y-, and Z-coordinates into a 2D array.
}
\examples{
# read landmark files
folder.with.landmarks <- ckpt2r_examples()
LM_list <- read_checkpoint(folder.with.landmarks)
                           
print(LM_list)

array_2D <- array_2D_from_df(df = LM_list, 
                specimen_column = "file_name",
                LM_column = "LM",
                X_column = "X",
                Y_column = "Y",
                Z_column = "Z")
                           
print(array_2D)

}
