% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{array_2D_from_LM_list}
\alias{array_2D_from_LM_list}
\title{Convert landmark list to table}
\usage{
array_2D_from_LM_list(LM.list, remove_NAs = FALSE, verbose = TRUE)
}
\arguments{
\item{remove_NAs}{remove landmarks that are missing in at least one specimen
from the whole dataset. Default: \code{TRUE}.}

\item{verbose}{If \code{TRUE} will inform the user if and which landmarks have
been removed in case \code{remove_NAs} was set to \code{TRUE}. Default: \code{TRUE}.}
}
\value{
The resulting data.frame will have the following dimensions:
\verb{n.specimens x n.landmarks*n.dimensions}. Rownames will be the filenames the
data came from.
}
\description{
Converts a landmark list loaded with \code{read_checkpoint()} to table (2D array).
}
\examples{
# read landmark files
folder.with.landmarks <- ckpt2r_examples()
LM.list <- read_checkpoint(folder.with.landmarks,
                           keep.missing = FALSE)
                           
print(LM.list)

# convert LM.list to 2D array
array_2D <- array_2D_from_LM_list(LM.list = LM.list)
dim(array_2D)

}
