% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ckpt2r_examples}
\alias{ckpt2r_examples}
\title{Get path to landmark example files}
\usage{
ckpt2r_examples()
}
\value{
Returns the path to the folder containing example Checkpoint files.
}
\description{
ckpt2r comes with example Checkpoint files. The files are stored in
\code{ckpt2r}'s \code{inst/extdata} folder, and this function returns the path to that
folder in examples.
}
