% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_checkpoint}
\alias{read_checkpoint}
\title{Read Checkpoint files from folder}
\usage{
read_checkpoint(folder, pattern = NULL, recursive = TRUE)
}
\arguments{
\item{folder}{A character string with the path to the Checkpoint files
(*.ckpt).}

\item{pattern}{Character string defining pattern that should be part of file
file names that the functions should consider. Default \code{NULL}.}

\item{recursive}{Logical value defininf if function should consider
subfolders (\code{TRUE}) or not (\code{FALSE}). Default: \code{FALSE}.}
}
\value{
Returns a tibble containing the following columns:\cr
\cr

| \strong{\code{file_name}} |   | name of the file that the list element was generated from |
| \strong{\code{LM}} |   | names of landmarks |
| \strong{\verb{X, Y, Z}} |  | x-, y- and z-coordinates of landmarks |
| \strong{\code{defined}} |   | missing (M) or present (N), as defined within
Checkpoint or (O) when not defined at all but present in other files.|
}
\description{
Allows reading Checkpoint files (*.ckpt) directly into R so that the
time-consuming export step within Checkpoint can be skipped.
}
\details{
So far, only point and curve landmarks are supported.
}
\examples{
# read landmark files
folder.with.landmarks <- ckpt2r_examples()
LM_list <- read_checkpoint(folder.with.landmarks)
                           
print(LM_list)

}
