% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_checkpoint}
\alias{read_checkpoint}
\title{Read Checkpoint files from folder}
\usage{
read_checkpoint(folder, pattern = NULL, recursive = FALSE, keep.missing = TRUE)
}
\arguments{
\item{folder}{A character string with the path to the Checkpoint files
(*.ckpt).}

\item{keep.missing}{If \code{FALSE}, landmarks marked as missing within
Checkpoint are removed, no matter if they are defined or not (i.e. no matter
if there are coordinate values present). Default: \code{FALSE}.}
}
\value{
Returns a list of which each list element consists of a dataframe
with the collowing columns:\cr
\cr

| \strong{\code{defined}} |   | missing (M) or present (N), as defined within Checkpoint. When keep.missing = FALSE, landmarks with the label M will be removed. |
| \strong{\verb{X, Y, Z}} |  | x-, y- and z-coordinates of landmarks |
| \strong{\code{LM}} |   | names of landmarks |
| \strong{\code{file}} |   | name of the file that the list element was generated from |
}
\description{
Allows reading Checkpoint files (*.ckpt) directly into R so that the
time-consuming export step within Checkpoint can be skipped.
}
\details{
So far, only point and curve landmarks are supported.
}
\examples{
# read landmark files
folder.with.landmarks <- ckpt2r_examples()
LM.list <- read_checkpoint(folder.with.landmarks,
                           keep.missing = FALSE)
                           
print(LM.list)

}
